refactor: 입찰자 수 자료형 변경

+++ src/main/java/com/example/budongbudong/domain/auction/dto/response/AuctionInfoResponse.java
@@ -5,7 +5,7 @@ import java.time.LocalDateTime;
 public record AuctionInfoResponse(
         Long startPrice,
         Long highestPrice,
-        Long totalBidders,
+        int totalBidders,
         LocalDateTime endedAt
 ) {
 }
